+++ src/armonik_cli/exceptions.py
@@ -1,0 +1,16 @@
+import rich_click as click
+
+
+class ArmoniKCLIError(click.ClickException):
+    """Base exception for ArmoniK CLI errors."""
+
+    def __init__(self, message: str) -> None:
+        super().__init__(message)
+
+
+class InternalError(ArmoniKCLIError):
+    """Error raised when an unknown internal error occured."""
+
+
+class NotFoundError(ArmoniKCLIError):
+    """Error raised when a given object of the API is not found."""
